fix migration and add data dummy seeder

+++ database/migrations/2019_06_24_064405_add_relationships_to_kapal_table.php
@@ -14,7 +14,7 @@ class AddRelationshipsToKapalTable extends Migration
     public function up()
     {
         Schema::table('kapal', function (Blueprint $table) {
-            $table->integer('id_maskapai')
+            $table->bigInteger('id_maskapai')
                   ->unsigned();
             $table->foreign('id_maskapai')
                   ->references('id')
